requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
chromedriver-binary==104.0.5112.20.0
cryptography==37.0.4
ffmpeg-python==0.2.0
future==0.18.2
h11==0.13.0
idna==3.3
iso8601==1.0.2
m3u8==2.0.0
outcome==1.2.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
requests==2.28.1
selenium==4.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.9
wsproto==1.1.0
